server/requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
supabase>=2.5.0
google-search-results>=2.4.2
pydantic>=2.8.0
httpx>=0.27.0
python-docx>=1.1.0
jinja2>=3.1.4